+++ src/commands/QueueController.php
@@ -15,7 +15,11 @@ use luya\admin\models\Config;
  * ```sh
  * ./vendor/bin/luya admin/queue
  * ```
- *
+ * 
+ * > As alternative to the admin/queue command you can enable {{luya\admin\Moudle::$autoBootstrapQueue}} "fake cronjob" behavior.
+ * 
+ * @see {{luya\admin\Moudle::$autoBootstrapQueue}}
+ * @author Basil Suter <basil@nadar.io>
  * @since 2.0.0
  */
 class QueueController extends Command