+++ src/ngrest/plugins/Select.php
@@ -15,9 +15,8 @@ use luya\admin\helpers\Angular;
 abstract class Select extends Plugin
 {
     /**
-     * If enabled the list tag will transform into an interactive schedling overlay.
-     *
-     * @var boolean
+     * @var boolean If enabled the list tag will transform into an interactive scheduling overlay. Keep in mind to turn on the queue system in order to enable time based scheduling
+     * @see {{luya\admin\Module::$autoBootstrapQueue}}
      * @since 2.0.0
      */
     public $scheduling = false;